Implement 'ignore' and 'fail' updaste-lock-response options
- With 'ignore' Fig will not be trying to create a lock.
- With 'fail' Fig will try to create a lock once and fail if not
  successful.
